Restore PNG output on Xcode build

+++ src/png_writer.cpp
@@ -40,7 +40,8 @@ png_writer::png_writer(std::string fn) {
 #endif
 }
 
-void png_writer::priv_write_header(size_t width, size_t height, size_t bpp) {
+void png_writer::priv_write_header(
+    unsigned int width, unsigned int height, unsigned int bpp) {
 #ifdef HAVE_LIBPNG
   png_set_IHDR(priv->png_ptr,
                priv->info_ptr,
